gozfs/holds: handle snapshot with no holds
zfs_get_holds is documented to fill in the return pointer on success. This does
not happen. Instead the list of holds is iteratively copied to the return list.
When there are no holds, the list is not touched. Passing in an all zero buffer
returns said all zero buffer, which is not a valid nvlist.

So to deal with this we copy an encoded empty list to the out buffer before
calling the ioctl. If there are no holds the same valid empty nvlist will be in
the returned buffer, otherwise the buffer will hold a new valid list.